refactor[#523] : widgets/follow/ui 타입 리팩토링
- FSD 구조를 맞추고 있지 않아 ui 폴더 생성
