+++ src/widgets/follow/ui/followNavigationBar.tsx
