+++ src/widgets/follow/ui/type.ts
@@ -1,0 +1,6 @@
+export interface FollowItemProps {
+  nickname: string;
+  petName: string;
+  profile: string | null;
+  isFollowing?: boolean;
+}
